chore: Split container media permutations page between side and top positions
